requirements.txt
python-hcl2

PyYAML~=6.0
